requirements.txt
streamlit
pandas
pymysql
sqlalchemy
python-dotenv
cryptography
tushare
chinese_calendar
mysql-connector-python
